Version 1
The Apk file is for installing in Android devices to play this RPG for
three kingdoms.
In this game, you're going to play the role of a newly come soldier in
Han dynasty of ancient China. You will be able to meet a lot of nobles
in china and record down your own history